WIP(obd) Sending VIN digits 1-by-1 looping twice for VIN (BROKEN)

+++ firmware_v5/telelogger/teleclient.h
@@ -27,7 +27,7 @@ public:
     void add(uint16_t pid, uint32_t value);
     void add(uint16_t pid, float value);
     void add(uint16_t pid, float value[]);
-    void add(uint16_t pid, int value[]);
+    void add(uint16_t pid, uint8_t value[]);
     void purge();
     void serialize(CStorage& store);
     uint32_t timestamp;
